Add encinitas-collector-go provisioning
* Objective

This commit objective is about completing the resources to provision
the encinitas-collector-go so it can be deployed in a server.

* Why

At some point, we will have to provision and deploy the
encinitas-collector-go, so these resources are a requirement for that
purpose.

* How

This commit completes the resources to provision the
encinitas-collector-go