Perform agent events & solana transactions processing
* Objective

This commit objective is about processing the agent events
and the Solana transactions to get their information ready
to be ingested.

* Why

Processing the incoming events is necessary to calculate
later and blend the agent and Solana information data.

* How

This commit treats the information from both sources and
uses Redis to store the initially processed information
that is ready to be transformed and ingested.