Retrieve peformance/throughput/apdex influx metrics
* Objective

This commit objective is about creating the endpoint that will retrieve
performance/throughput/apdex metrics from the influx repository.

* Why

This endpoint is necessary for the encinitas-ui to craft the dashboard
performance/throughput/apdex charts.

* How

This commit adds the new metrics endpoint, and craft the required
repository methods to perform the operations.