Produce semi-mocked data for the dashboard's metrics
* Objective

This commit objective is about producing some mix of natural/mocked data
for the dashboard metrics, performance, apdex, throughput, and error
rate.

* Why

We can not provide "real data" yet because we have no proper validator
configured, and because we have no adequate integration in place, both
things are necessary for encinitas to operate.

* How

This commit is fetching actual transaction data from a Pyth Oracle in
testnet but then generates some demo data for some fields in order
to properly return dashboard data.