Add GET Transactions repositories & endpoint
* Objective

This commit objective is about adding the GET transaction endpoint which
would return all the transactions and their percentage in load terms
overall.

* Why

This endpoint is meant to feed the Transactions Page in the
encinitas-ui.

* How

This commit introduces a new Solana Postgres SQL method that performs
the fetching of the transaction information and creates a Handler that
uses that endpoint