Update docker-compose setup
* Objective

This commit objective is about simplying the docker-compose setup to use
the local resources that has been provisioned instead of local images.

* Why

In this phase for encinitas this is a better option because we want to
process both locally and remotelly the same set of transactions, so they
are available for development but also for the alpha. This is not a good
practice but is just for the alpha.

* How

This commit removes the docker dependencies of influx and redis.