Add Program performance & throughput
* Objective

This commit objective is to add the Program performance and throughput
resources needed to create the visualizations for the details of the
given Program executions.

* Why

The transactions details page is yet the second most important page for
Encinitas, and this commit adds the endpoints that the page needs to
render.

* How

This commit adds the endpoint and repository methods for the Performance
and Throughput program metrics