Get influx error from the error channel
* Objective

This commit objective is about capturing the influx's errors on the writer
by initializing their capture while initializing an influx's repository.

* Why

We are noticing that the influx write operations are not working
properly, it's not writing anything on the database... this is
an attempt to capture what's going on.

* How

This commit initializes a go routing to capture the influx error
channel.