Add emails repository
* Objective

This commit objective is about adding the email repository so we can
start collecting emails for the Alpha.

* Why

We need to get people to use the App, and we will do an initial launch for
the website where there will be a landing page, some information about how
the integration will be, and a link for the example.

* How

This commit adds the repository, the service, and the endpoint that will
be used to register emails