Use Telefrag for writing in influxdb
* Objective

This commit objective is about changing the influx repository so it uses
Telegraf instead of writing directly onto influxdb.

* Why

The main reason is: performance; it has been a nightmare to get influxdb
to behave, likely because of the way that we are writing metrics, etc.

* How

This commit changes the influx repository to use the telegraf agent
instead of writing directly to influxdb