Reduce metric timespan to 6h
* Objective

This commit obejctive is about reducing the span for the metrics to 6h.

* Why

We had a downtime for the validator service that caused a gap on the
metrics so this commit intends to make it better for the alpha.

* How

This commit reduces the metric span to 6h from the previously configured
2 days